Correct interrupt priorities.
There are 8 supported priorities on the NXP43xx M4 core, from
0 to 224 in steps of 32. Lower values are higher priority. The
lowest 5 bits are ignored, so writing 255 sets priority 224.

Set the M0 event interrupt to the lowest priority, and raise the
USB0 interrupt priority from the lowest to the second lowest.

This prevents the M0 events from interrupting the USB0 ISR.